Refactor: Merge EC and RSA key pair repositories
Merged EcKeyPairRepository and RsaKeyPairRepository into a single KeyPairRepository interface, along with their implementations.
This includes updating KeyAttestationViewModel to use the unified repository and ensuring Hilt DI provides the correct dependencies.
All temporary developer comments added during this refactoring process have been removed.
